adding delete function to remove files on the cloud, fixed changes so that you can run as day 1.1, etc.

+++ webInterface/rtAtten/RtAttenWeb.py
@@ -446,7 +446,7 @@ class RtAttenWeb():
         try:
             regConfig = request['regConfig']
             regType = request['regType']
-            dayNum = int(regConfig['dayNum'])
+            dayNum = float(regConfig['dayNum'])
         except KeyError as err:
             RtAttenWeb.webServer.setUserError("Registration missing a parameter ('regConfig', 'regType', 'dayNum')")
             return
